- Add projects argument under browse-osp
Used urllib to extract easily the project names in every URL provided by the  section of the packages.
In this way we can parse the Git URL (extracted with the net_loc) and receive the projecs (Which is the path) following the [URL Syntactic Components](https://www.rfc-editor.org/rfc/rfc1808.html#section-2.1).