Introduce exclude mechanism for generated jobs
This change allows users of the project to easily
update a dictionary based map in order to affect which
jobs will be included in the generated templates.

This can be controlled in different levels:

* Job level: exclude job no matter which tag or project
             are being used
* Tag level: exclude jobs when a specific tag is being used
* Project level: exclude jobs for specific projects

This granularity will help us to have better control
at what exactly is included in the generated templates
and for the user this is a simple structure to update in
one place.

Change-Id: I02ce21ae3eab06140dfb64238f1485cb9b5e34c3