Notify user if a component doesn't exists
Before this change, if user used a component which doesn't exists,
the program would exit without an error and the user didn't have any
indication whether he/she even specified a valid component.

With this chain, user will be notified if the component he/she specified
with --component doesn't exists and the app will also list available
components. This should ease the use of the browser sub-command.

Change-Id: Icfd904936cdef085a5a7f8580e656cbd488c2d22