Fetch git history in CI
This commit enables fetching git commits history in our CI.
By default the alpine image has no git client and therefore
the checkout action fetches sources from a tarball, not by
cloning the repository.

However, in `tox` recently I enabled installing the Znoyder
into the virtual environment and because of that it complains
about missing data for versioning (for now we rely on counting
the number of commits).