Introduce jobs mapper
This commit introduces the jobs mapper module and migrates
existing code into it. The module will take care of translating
defined jobs mapping and will make sure it has the desired effect.

As part of the introduction, the existing exclude and include
maps were translated to the new fashion, with necessary code
changes around (e.g. tests).

In the future the additional override and copy mechanisms
will be implemented as well.

Jira: OSPCRE-208