Jobs mapper improvements
This commit introduces changes to jobs mapper that does not
replace all jobs options in override map, but only updates
the defined values and allows us to undefine some jobs options
we may not want to keep by setting them to None (tilde ~ in YAML).
Additional unit tests were introduced to ensure the mechanism
works as expected.

This commit also solved one identified flaw in Zuul library
that caused unexpected behavior due to mess in objects references.
Further investigation and improvements to Zuul library will be
implemented in the future.