Reorganize tests
This commit moves existing unit tests under the znoyder/ directory,
so it is possible to run only specific test case or a single test,
which is very useful during debugging.

Example: tox -e unit -- znoyder.tests.module.TestCase.test_method

Additionally the tox environment were reorganized a little and
the yamllint environment was introduced to ensure the sanity
of Znoyder config map file. Similarly the doctests environment
was introduced to find and verify and compatible docstring.