Improve skipping in downloader
Now the downloaded will always create a target directory
for a given repository and will skip immediately when
such directory already exists. This allows us to limit
requests in local environment when testing changes.

Without this commit, for projects that have no Zuul
configuration, we would always request the repository.
In local environment, when testing the configs generation,
we can easily get blocked by GitHub API due to that.

It should not affect the first run in clean environment,
such as in Jenkins job.