Update jobs names in config maps
We genrate the downstream config, for downstream project names,
so we should rely on names that we have in downstream. Before
this commit, the project names were coming from downstream,
but defined jobs names were corresponding to upstream names.
With the templater refactor, it will be all unified to downstream
names everywhere.