Refactor generator code
This commit introduces a major refactor of the configs generator.
The big main function was split into smaller ones for clearness.
The logic of pipelines preparation was moved from the templater,
which as side effect allowed to introduce easily to support
multiple OSP releases during generation.

The generator now produces two directories with results:
– osp-internal-jobs – that contains the projects templates,
– osp-internal-jobs-config – that contains the projects config.

Although many parts of the original code were reused, for review
it is better to just view the new code than compare the diffs,
as this component was basically reinvented again.