Remove old generator and templater tests
The existing tests were just drafts, as we already knew for a long
time that the components will be refactored. As these components
were mostly reinvented, it will be easier to write tests from scratch,
so in this commit I am simply removing the existing tests
for CI to pass during pull request and first major review.