Move from tox-py39 to rpm-py39 jobs
This commit changes the existing unit tests jobs
to the newly created RPM-jobs™, which provides
additional benefit of testing the actual Python
dependencies that are shipped as part of OSP product.

Temporary config map with non-votings is applied
for the initial triaging phase to the jobs
that currently fail for some reason.