Prevent jobs duplicates during include
For some upstream project, there is a situation where the same
job is defined in multiple places, e.g. in a project-template
and a project itself. Our finder implementation in such case
included the job twice, which was incorrect behavior – the latter
definition in the tree (template -> template -> project) should
always overwrite the previous definition in such case.
This commit introduces a solution for mentioned problem
by involving a set object and traversing the list of templates
in reversed order. The hash function for a ZuulJob object
is defined over a pair of job name and job pipeline,
so it properly covers the shadowing of previous definitions,
at least according to conducted tests ;-)