Add initial jobs OSP-17.1
This commit utilizes the copy mechanism of Znoyder and the amazing
shadowing effect in YAML files to generate initial list of jobs
for OSP 17.1 release. Basically, all jobs are copied under the same
name, but with additional `branches` parameter, which overrides
the default value added there from template.

This is not final solution, but for now, as we work on stabilizing
the jobs for OSP 17.0, it would provide us with initial results.
Then, the proper including, adding and overriding will be implemented
instead of the copy map entries.