Refactor logs suppression
This commit changes the logs suppression mechanism to work
per Znoyder tests module. This is because the previous solution
changed the setting globally for all tests, which could lead
to unpredicted behaviors, depending on the tests execution order.
Here the logs are disabled on demand for specific tests modules
and restored after those modules are processed.