Fix for nova
There is a unit test that does not mock libvirt/libguestfs properly
and when those are installed in system, there goes a real call
which results with timeout due to no expected response.
The test will be fixed in upstream, but for now we just remove
the problematic packages that come from spec file.