Inherit py38 as py39 jobs instead of py39 only
I noticed that there are few projects in upstream that
contain definitions for py38, but not for py39 (which is
expected, as py38 was official for stable/wallaby).
Hence the proposal is to take py38 jobs and rename to py39,
which extends the downstream coverage to following projects:
- networking-l2gw
- python-designateclient
- python-keystoneauth1
- python-magnumclient
- python-osc-placement
- python-proliantutils
- python-tooz
The only drawback is that scciclient has no jobs for py38,
so we replace openstacksdk with scciclient in add map
to still contain jobs for this project.