Add OSP 18 jobs for Nova
We're slowly starting to see OSP 18.0 backports to Nova downstream.
Add basic functional and unit tests jobs as a sanity check for those
backports. These are copied verbatim from the 17.1 versions (minus the
py36 unit tests job), working on the assumption that we can run them
on RHEL 9.2 for now and everything will be fine.