Do not install python3-mock and python3-libvirt
Nova's code mocks out compute manager's with a mock.DEFAULT.
However, the python3-placement-tests requires the python3-mock
package – we should remove it to use the built-in standard library.

Then there was error regarding python3-libvirt module imported,
while it should not be installed. Hence we remove the package.
This resolves the issues with osp-rpm-functional-py39 job.