More fixes for OSP 18 projects
This includes fixes for cinder, heat, oslo.rootwrap and aodhclient.
There were AttributeErrors in PyOpenSSL:
> module 'lib' has no attribute 'X509_V_FLAG_CB_ISSUER_CHECK'
Also for projects that decided to switch to tox4, the --sitepackages
option was deprecated in favor of VIRTUALENV_SYSTEM_SITE_PACKAGES=1.
The tests suite for oslo.rootwrap invokes a new process and in case
of tox with sitepackages it misses the venv-installed packages,
hence the project is installed out of venv as well.