Apply proper upper-contraints file to neutron OSP17.1 pep8 job
The OSP17.1 neutron version requires a new version of neutron-lib.
When pep8 is executed, it fails because the upstream wallaby
constrainsts file sets neutron-lib version to 2.10.2.
This patch fixes the problem by using an alternative file that sets
neutron-lib version to 2.10.3.
NOTE: the wallaby constraints file will not change anymore at this
stage of the project.