Add selective cache option
This commit customizes the cache decorator so it is possible
to specify what parameters will be used as cache keys ~ i.e.
for function that we actually care about selected parameters that
makes the actual difference we can consider only specified ones
(by default all function parameters are considered).

Additionally it is possible to produce a human-readable string
instead of a calculated hash – this would make the generated
cache file browsable for us in case we need that...