Make cache for generator predictable
This commit adds sorting of ZuulJobs entries so when cached
the produced output is always predictable and repeatable
across multiple calls to generator, no matter who, where
and when executes the code. This should allow us to include
small incremental changes in cache file when regenerating
the cache periodically for any upstream changes.