Fix get_packages by osp-project to cover non-url strings
Currently, the urlparse lib is used to find the short repo name from the
osp-patches string in distroinfo. This does not work for non-url based
strings as `git@gitlab.com:eng/test/repo_name.git`.

This patch is moving to use regex instead of urlparse asuming that the
repo name is the string after the last `/` and removing the `.git` if
exists. This should cover all kind of strings that may be used in
osp-patches field.

Note this is changing slightly the behavior for repos in namespaces.
i.e. before this patch if osp-patches is `ssh://server.com/foo/bar`,
osp-project would be `foo/bar`. After this patch, it is `bar`. This
behavior matches the current usage of this parameter which is used to
matcht zuul.project.short_name which points always to the last part,
`bar` in the previous case. This has not been affecting in the past
because the osp-patches used were not namespaced.