Use basename instead of regex
It turns out that in some cases we have projects
in ospinfo that have no downstream forks – in such cases,
there is osp-patches key set to empty string and regular
expression breaks. Also, the solution with regular expression
may lead to ambiguity (is a/b/c matched as a+b/c or a/b+c?),
hence we replace that with just basename from path.